Added __attribute__((printf())) to debugf() and fixed format string errors.

+++ src/fsexaminenext.c
@@ -29,7 +29,7 @@ static int Fbx_releasedir(struct FbxFS *fs, const char *path, struct fuse_file_i
 static int Fbx_readdir(struct FbxFS *fs, const char *path, APTR udata, fuse_fill_dir_t func,
 	QUAD offset, struct fuse_file_info *fi)
 {
-	ODEBUGF("Fbx_readdir(%p, '%s', %p, %p, %lld, %p)\n", fs, path, udata, func, offset, fi);
+	ODEBUGF("Fbx_readdir(%p, '%s', %p, %p, %lld, %p)\n", fs, path, udata, func, (long long)offset, fi);
 
 	return FSOP readdir(path, udata, func, offset, fi, &fs->fcntx);
 }
